requirements.txt
Django
Pillow
numpy
matplotlib
opencv-python